Tests separated in different files
